refactor: move providers and style file out of app directory
